Move `get_seeds_per_chain` and Random type-hint variables to util.py
